+++ test/plugin/store_dimensions_test.rb
@@ -127,6 +127,10 @@ describe Shrine::Plugins::StoreDimensions do
     assert_equal 0, io.pos
   end
 
+  it "has .dimensions alias" do
+    assert_equal [100, 67], @shrine.dimensions(image)
+  end
+
   it "returns Shrine::Error on unknown analyzer" do
     assert_raises Shrine::Error do
       @shrine.plugin :store_dimensions, analyzer: :foo